+++ database/migrations/2024_02_0001_V_1_0/2024_02_02_205906_create_stock_has_material_distributions_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
 
-            $table->integer('equipment_id')->unsigned();
-            $table->integer('stock_id')->unsigned();
+            $table->bigInteger('equipment_id')->unsigned();
+            $table->bigInteger('stock_id')->unsigned();
             $table->integer('percent_profit')->unsigned();
             $table->float('price_reference', 10, 2, true);
